\name{adjust_X_chr}
\alias{adjust_X_chr}
\title{adjust_X_chr}
\usage{
adjust_X_chr(ff_list, X_chr, Xfactor)
}
\arguments{
  \item{ff_list}{list of ff objects each representing data
  of one chromosome}

  \item{X_chr}{name of the X chromosome in ff_list}

  \item{Xfactor}{factor by which the counts should be
  adjusted (2 for male celllines, 4/3 for male/female
  mixtures)}
}
\value{
adjusted ff_list
}
\description{
Adjusts lower X chromosome counts due cells being male or a
mixture of male/female.
}
\author{
Mark Heron
}

