% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_represenation.R
\name{evenfasta2num}
\alias{evenfasta2num}
\title{evenfasta2num}
\usage{
evenfasta2num(fastas)
}
\arguments{
\item{fastas}{(DNAStringSet) that can be different length DNA sequences that should be converted (single DNAString also possible)}
}
\value{
list of numeric representation of the DNA sequences
}
\description{
Transforms a DNAStringSet into a list of numeric representation of the mono-nucleotides.
Any mono-nucleotide that isn't A,C,G or T is represented as **NA**.
This is a streamlined version of \code{\link{fasta2num}} for mono-nucleotides that additionally allows varying sequence lengths.
}

