\name{get_dyad_pos}
\alias{get_dyad_pos}
\title{get_dyad_pos}
\usage{
get_dyad_pos(data_list, dyad_base = "center", offset = 73)
}
\arguments{
  \item{data_list}{list of mapped fragments for each
  chromosome}

  \item{dyad_base}{based on what position should the dyad
  position be calculated}

  \item{offset}{offset if the dyad position isn't
  calculated from the center}
}
\value{
list of ff matricies with dyad positions and intensity
}
\description{
Extracts the dyad positions in the chosen way from a table
with mapped fragment start and ends
}
\author{
Mark Heron
}

