% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_handling.R
\name{getRunningWindowCG}
\alias{getRunningWindowCG}
\title{getRunningWindowCG}
\usage{
getRunningWindowCG(fastas, half_window_size = 73)
}
\arguments{
\item{fastas}{(DNAStringSet) for which to calculate the running window C+G\%}

\item{half_window_size}{how much to extend the window in either direction}
}
\value{
list of ff vectors that contain the running C+G\%
}
\description{
Computes a running window C+G\% for a DNAStringSet.
}

