% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_represenation.R
\name{num2freq}
\alias{num2freq}
\title{num2freq}
\usage{
num2freq(seqnum, oligo_length)
}
\arguments{
\item{seqnum}{(numeric matrix) numerically encoded oligo-nucleotide representation of DNA sequences}

\item{oligo_length}{(integer) length of the oligonucleotide-coding used (order+1)}
}
\value{
matrix of oligonucleotide frequencies
}
\description{
Calculates the oligo-nucleotide frequencies given a numerically encoded matrix representation.
}

