% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mers.R
\name{oligo_names}
\alias{oligo_names}
\title{create oligo-nucleotide names}
\usage{
oligo_names(oligo_length)
}
\arguments{
\item{oligo_length}{(numeric) length of the oligo-nucleotides to use.}
}
\value{
(character vector) oligo-nucleotide name vector
}
\description{
Creates an array with all possible oligo-nucleotides of the length \code{mer_length}, sorted alphabetically.
}
\examples{
oligo_names(2)

}
