% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logos.R
\name{pwm2ic}
\alias{pwm2ic}
\title{PWM information content}
\usage{
pwm2ic(pwm)
}
\arguments{
\item{pwm}{(numeric matrix) position weight matrix}
}
\value{
(numeric vector) of the information content per position
}
\description{
Compute the information content per position for a PWM.
}
