% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logos.R
\name{pwm2ig}
\alias{pwm2ig}
\title{PWM information gain}
\usage{
pwm2ig(pwm, bg)
}
\arguments{
\item{pwm}{(numeric matrix) position weight matrix}

\item{bg}{(numeric vector) background frequencies of the nucleotides}
}
\value{
(numeric vector) of the information content per position
}
\description{
Compute the information gain per position for a PWM.
}
