% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logos.R
\name{revComp}
\alias{revComp}
\title{Reverse complement a PWM}
\usage{
revComp(pwm)
}
\arguments{
\item{pwm}{(numeric matrix) position weight matrix}
}
\value{
(numeric matrix) reverse complement position weight matrix
}
\description{
Create the reverse complement PWM
}

