% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mers.R
\name{complementary_oligo_positions}
\alias{complementary_oligo_positions}
\title{complementary_oligo_positions}
\usage{
complementary_oligo_positions(oligo_length)
}
\arguments{
\item{oligo_length}{(numeric) length of the oligo-nucleotides to use.}
}
\value{
position of the matching reverse complement oligo-nucleotides
}
\description{
complementary_oligo_positions
}
\examples{
complementary_oligo_positions(2)
}

