% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_handling.R
\name{convertSparse2Complete_ff}
\alias{convertSparse2Complete_ff}
\title{convertSparse2Complete_ff}
\usage{
convertSparse2Complete_ff(sparse, lengths)
}
\arguments{
\item{sparse}{list of sparse representation matricies}

\item{lengths}{list of vector lengths, names must match those of \code{sparse}}
}
\value{
list of ff vectors, which are the complete representation of each of \code{sparse} elements
}
\description{
Converts sparse representation to complete ff vector representation for a list of data.
}
