% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_boolean_representation.R
\name{fasta2sparse_boolean}
\alias{fasta2sparse_boolean}
\title{fasta2sparse_boolean}
\usage{
fasta2sparse_boolean(fasta, oligo_length)
}
\arguments{
\item{fasta}{(DNAString) sequence to encode}

\item{oligo_length}{(integer) oligo-nucleotide length to encode the sequence in}
}
\value{
(boolean matrix) sparse encoding of the DNAString, columns=positions, rows=oligo-nucleotides
}
\description{
Creates sparse encoded matrix of a DNA sequence.
}
