% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_boolean_representation.R
\name{fasta2sparse_boolean_ff}
\alias{fasta2sparse_boolean_ff}
\title{fasta2sparse_boolean_ff}
\usage{
fasta2sparse_boolean_ff(fasta, oligo_length)
}
\arguments{
\item{fasta}{DNAString}

\item{oligo_length}{oligonucleotide length to code the sequence in}
}
\value{
sparse ff.matrix representation of the DNAString
}
\description{
Creates sparse coding ff object of a fasta sequence.
}
\author{
Mark Heron
}

