\name{num2freq}
\alias{num2freq}
\title{num2freq}
\usage{
num2freq(seqnum, oligo_length)
}
\arguments{
  \item{seqnum}{(numeric matrix) oligo-nucleotide-coded
  representation of fasta sequences}

  \item{oligo_length}{(integer) length of the
  oligonucleotide-coding used (order+1)}
}
\value{
matrix of oligonucleotide frequencies
}
\description{
Calculates the mere frequencies given a
oligo-nucleotide-coded representation matrix
}

