% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_represenation.R
\name{num2weightedfreq}
\alias{num2weightedfreq}
\title{num2weightedfreq}
\usage{
num2weightedfreq(seqnum, weights, oligo_length)
}
\arguments{
\item{seqnum}{(numeric matrix) numerically encoded oligo-nucleotide representation of DNA sequences}

\item{weights}{(numeric) to be given to the fasta sequences (should have length matching \code{ncol(seqnum)})}

\item{oligo_length}{(integer) length of the oligonucleotide-coding used (order+1)}
}
\value{
matrix of weighted oligonucleotide frequencies
}
\description{
Calculates the weighted oligo-nucleotide frequencies given a numerically encoded matrix representation.
}

