% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotOligoFreqs}
\alias{plotOligoFreqs}
\title{plotOligoFreqs}
\usage{
plotOligoFreqs(freqs, x_pos = NULL, ylim = NULL,
  xlab = "distance to position of interest", ylab = "frequencies",
  legend_nrow = 1, col = NULL, ...)
}
\arguments{
\item{freqs}{(numeric matrix) oligonucleotide frequencies}

\item{x_pos}{(numeric vector) x axis positions}

\item{ylim}{(numeric tuple) range of the y axis}

\item{xlab}{(character) x axis label, defaults to: \"distance to position of interest\"}

\item{legend_nrow}{(integer) over how many rows should the legend be split}

\item{col}{(character vector) colours to use, if NULL (default) a palette will be generated by distinctive_colours()}

\item{...}{further parameters for \code{\link{plot}}}
}
\description{
plots oligonucleotide frequencies over a region as lines
}

