% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pRon.R
\docType{package}
\name{pRon-package}
\alias{pRon-package}
\title{Pretty representations of oligo-nucleotides}
\description{
A selection functions to help handle oligo-nucleotides (mers), including creation frequency/energy profile figures.
}
\author{
Mark Heron
}

