% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_handling.R
\name{read_genome_fasta}
\alias{read_genome_fasta}
\title{read_genome_fasta}
\usage{
read_genome_fasta(genome_dir)
}
\arguments{
\item{genome_dir}{the directory where the chromosome fasta files can be found}
}
\value{
DNAStringSet of the genome
}
\description{
Load chromosome fasta files from a genome folder into a DNAStringSet object.
}

