% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logos.R
\name{seqLogo}
\alias{seqLogo}
\title{seqLogo}
\usage{
seqLogo(pwm, background = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25),
  ytype = "information", xaxis = TRUE, xlim = c(0, ncol(pwm)),
  yaxis = TRUE, ylim = NULL, xlab = "Position", xfontsize = 15,
  yfontsize = 15, region = c(0, 0, 1, 1), ...)
}
\arguments{
\item{pwm}{position weight matrix to create a Sequence Logo from}

\item{background}{(numeric vector-4) background probabilities of the nucleotides for information gain computation default all=0.25}

\item{ytype}{(character) the type of y-axis to be used "information" (or "ic", "content") for information content (default); "relative" (or "gain") for information gain; "probabilities" (or anything else) for probabilities}

\item{xaxis}{(logical) should an x-axis be drawn or vector of labels for the xaxis (must be same length as pwm)}

\item{xlim}{(numeric vector-2) default c(0,ncol(pwm)) see graphical parameters}

\item{yaxis}{(logical) should an y-axis be drawn}

\item{ylim}{(numeric vector-2) default NULL see graphical parameters}

\item{xlab}{(character) see graphical parameters}

\item{xfontsize}{(integer) size for x-axis text}

\item{yfontsize}{(integer) size for y-axis text}

\item{region}{in which to plot the figure, set by four margin values, set to \code{par("mar")} for default full plot}

\item{...}{further parameters to be passed through to dataViewport}
}
\description{
Plot a PWM sequence logo.
}
\examples{
pwm <- rbind( A=c( 0,  0, 15,  5,  5,  1, 0.1),
              C=c(10,  0,  0,  5,  5,  2, 0.1),
              G=c(10,  0,  5,  0,  5,  3, 0.1),
              T=c( 0, 20,  0, 10,  5, 14,19.7)) /20
              
seqLogo(pwm, region=par("mar"))

}
